+++ apps/app/src/components/order/PaymentPendingWarning.tsx
@@ -1,7 +1,7 @@
 import { NavigationProp, useNavigation } from '@react-navigation/native';
 import { OrderStatus } from '../../data/types';
 import { useUpdateOrderMutation } from '../../data/mutations';
-import { AppStackParamList } from '../../types/navigation';
+import { AppStackParamList } from '../../navigation/types';
 import { StyleSheet, View } from 'react-native';
 import { Button, Spacer, Typography, useTheme } from '@habiti/components';
