+++ apps/app/src/screens/Verify.tsx
@@ -1,53 +1,92 @@
-import { Button, Screen, Spacer, Typography } from '@habiti/components';
-import { RouteProp, useRoute } from '@react-navigation/native';
 import React from 'react';
-import { StyleSheet, TextInput, View } from 'react-native';
+import { Pressable, View, StyleSheet } from 'react-native';
+import {
+	Button,
+	FormInput,
+	Icon,
+	Screen,
+	Spacer,
+	Typography
+} from '@habiti/components';
+import { zodResolver } from '@hookform/resolvers/zod';
+import { RouteProp, useRoute } from '@react-navigation/native';
+import { useForm } from 'react-hook-form';
+import { SafeAreaView } from 'react-native-safe-area-context';
+import { z } from 'zod';
 
-import CodeInput from '../components/verify/CodeInput';
 import { useVerifyCodeMutation } from '../hooks/mutations';
-import { AppStackParamList } from '../types/navigation';
+import { AppStackParamList, AppStackScreenProps } from '../navigation/types';
+import CodeInput from '../components/verify/CodeInput';
 
-const VerifyAuthentication = () => {
+const verifySchema = z.object({
+	code: z
+		.string()
+		.length(6)
+		.regex(/^\d{6}$/)
+});
+
+type VerifyFormValues = z.infer<typeof verifySchema>;
+
+const Verify = ({ navigation }: AppStackScreenProps<'Verify'>) => {
 	const { params } = useRoute<RouteProp<AppStackParamList, 'Verify'>>();
-	const [code, setCode] = React.useState('');
 	const verifyCodeMutation = useVerifyCodeMutation();
 
-	const handleSubmit = () => {
+	const methods = useForm<VerifyFormValues>({
+		resolver: zodResolver(verifySchema),
+		defaultValues: { code: '' },
+		mode: 'onChange'
+	});
+
+	const code = methods.watch('code');
+
+	const onSubmit = (values: VerifyFormValues) => {
 		verifyCodeMutation.mutate({
 			email: params.email,
-			code
+			code: values.code
 		});
 	};
 
+	const handleBack = () => {
+		navigation.goBack();
+	};
+
 	return (
-		<Screen style={{ justifyContent: 'center', paddingHorizontal: 16 }}>
-			<Typography size='xxxlarge' weight='bold'>
-				Enter verification code
-			</Typography>
-			<Typography variant='secondary'>
-				A verification code was sent to your email.
-			</Typography>
-			<TextInput
-				autoFocus
-				style={styles.hidden}
-				onChangeText={setCode}
-				keyboardType='number-pad'
-				maxLength={6}
-			/>
-			<Spacer y={16} />
-			<View style={styles.inputs}>
-				{Array(6)
-					.fill(0)
-					.map((_, index) => (
-						<CodeInput key={index} value={code[index]} />
-					))}
-			</View>
-			<Spacer y={32} />
-			<Button
-				text='Verify'
-				onPress={handleSubmit}
-				loading={verifyCodeMutation.isPending}
-			/>
+		<Screen>
+			<SafeAreaView>
+				<Pressable onPress={handleBack}>
+					<Icon name='chevron-left' />
+				</Pressable>
+				<Spacer y={16} />
+				<Typography size='xxxlarge' weight='bold'>
+					Enter verification code
+				</Typography>
+				<Typography variant='secondary'>
+					A verification code was sent to {params.email}
+				</Typography>
+				<FormInput
+					name='code'
+					control={methods.control}
+					autoFocus
+					style={styles.hidden}
+					keyboardType='number-pad'
+					maxLength={6}
+				/>
+				<Spacer y={16} />
+				<View style={styles.inputs}>
+					{Array(6)
+						.fill(0)
+						.map((_, index) => (
+							<CodeInput key={index} value={code[index]} />
+						))}
+				</View>
+				<Spacer y={32} />
+				<Button
+					text='Verify'
+					onPress={methods.handleSubmit(onSubmit)}
+					loading={verifyCodeMutation.isPending}
+					disabled={!methods.formState.isValid}
+				/>
+			</SafeAreaView>
 		</Screen>
 	);
 };
@@ -66,4 +105,4 @@ const styles = StyleSheet.create({
 	}
 });
 
-export default VerifyAuthentication;
+export default Verify;
